requirements.txt
Django==1.11.5
django-appconf==1.0.2
django-avatar==4.0.1
django-bootstrap3==9.0.0
django-braces==1.11.0
django-model-utils==3.0.0
django-notifications-hq==1.3
django-registration-redux==1.7
django-smartfields==1.0.9
jsonfield==2.0.2
markdown2==2.3.4
olefile==0.44
Pillow==4.2.1
pytz==2017.2
